Desarrollo/requirements.txt
# Framework web
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Base de datos
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Autenticación y seguridad
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# Machine Learning
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
tensorflow==2.15.0
joblib==1.3.2

# Procesamiento de lenguaje natural
google-generativeai==0.3.2
openai==1.12.0

# Utilidades
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
tenacity==8.2.3
pytz==2024.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
pytest-cov==4.1.0

# Documentación
mkdocs==1.5.3
mkdocs-material==9.5.3

# Logging y monitoreo
structlog==24.1.0
prometheus-client==0.19.0

# Additional dependencies
lightgbm>=4.3.0    # O xgboost, según el modelo usado
python-dateutil>=2.8.2
dill>=0.3.8 